scripts/agent/requirements.txt
anthropic>=0.40.0
supabase>=2.0.0
httpx>=0.27.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
